refactor: widget components (data-comp, chart-tabs)
