+++ static_report/src/components/shared/NumericTableColumn.tsx
@@ -4,7 +4,10 @@ import { ZColSchema, zReport } from '../../types';
 import {
   formatAsAbbreviatedNumber,
   formatColumnValueWith,
+  formatIntervalMinMax,
+  formatNumber,
 } from '../../utils/formatters';
+import { getColumnDetails } from '../../utils/transformers';
 import { NO_VALUE } from './ColumnCard/ColumnTypeDetail/constants';
 import { MetricsInfo } from './MetricsInfo';
 
@@ -22,179 +25,138 @@ export function NumericTableColumn({
   const isTargetUndefined = targetColumn === undefined;
   zReport(ZColSchema.safeParse(baseColumn));
   zReport(ZColSchema.safeParse(baseColumn));
+  if (baseColumn) {
+    zReport(ZColSchema.safeParse(baseColumn));
+    var { distinct: baseDistinct, duplicates: baseDuplicates } = baseColumn;
+    var {
+      distinctOfTotal: baseDistinctOfTotal,
+      duplicatesOfTotal: baseDuplicatesOfTotal,
+    } = getColumnDetails(baseColumn);
+  }
+
+  if (targetColumn) {
+    ZColSchema.parse(targetColumn);
+    var {
+      distinctOfTotal: targetDistinctOfTotal,
+      duplicatesOfTotal: targetDuplicatesOfTotal,
+    } = getColumnDetails(targetColumn);
+  }
   return (
     <>
       <Flex direction="column">
-        <MetricsInfo
-          {...props}
-          name="Average"
-          metakey="avg"
-          firstSlot={formatColumnValueWith(
-            baseColumn?.avg,
-            formatAsAbbreviatedNumber,
-          )}
-          secondSlot={
-            isTargetNull
-              ? NO_VALUE
-              : isTargetUndefined
-              ? undefined
-              : formatColumnValueWith(
-                  targetColumn?.avg,
-                  formatAsAbbreviatedNumber,
-                )
-          }
-        />
-        <MetricsInfo
-          {...props}
-          name="SD"
-          metakey="stddev"
-          firstSlot={formatColumnValueWith(
-            baseColumn?.stddev,
-            formatAsAbbreviatedNumber,
-          )}
-          secondSlot={
-            isTargetNull
-              ? NO_VALUE
-              : isTargetUndefined
-              ? undefined
-              : formatColumnValueWith(
-                  targetColumn?.stddev,
-                  formatAsAbbreviatedNumber,
-                )
-          }
-        />
+        {baseColumn?.type !== 'datetime' && baseColumn?.type !== 'other' && (
+          <>
+            <MetricsInfo
+              {...props}
+              name="Average"
+              metakey="avg"
+              firstSlot={formatColumnValueWith(
+                baseColumn?.avg,
+                formatAsAbbreviatedNumber,
+              )}
+              secondSlot={
+                isTargetNull
+                  ? NO_VALUE
+                  : isTargetUndefined
+                  ? undefined
+                  : formatColumnValueWith(
+                      targetColumn?.avg,
+                      formatAsAbbreviatedNumber,
+                    )
+              }
+            />
+            <MetricsInfo
+              {...props}
+              name="SD"
+              metakey="stddev"
+              firstSlot={formatColumnValueWith(
+                baseColumn?.stddev,
+                formatAsAbbreviatedNumber,
+              )}
+              secondSlot={
+                isTargetNull
+                  ? NO_VALUE
+                  : isTargetUndefined
+                  ? undefined
+                  : formatColumnValueWith(
+                      targetColumn?.stddev,
+                      formatAsAbbreviatedNumber,
+                    )
+              }
+            />
+          </>
+        )}
       </Flex>
       <Flex direction="column">
+        {(baseColumn?.type === 'numeric' || baseColumn?.type === 'integer') && (
+          <>
+            <MetricsInfo
+              {...props}
+              name="Min"
+              metakey="min"
+              firstSlot={formatColumnValueWith(
+                baseColumn?.min,
+                formatAsAbbreviatedNumber,
+              )}
+              secondSlot={
+                isTargetNull
+                  ? NO_VALUE
+                  : isTargetUndefined
+                  ? undefined
+                  : formatColumnValueWith(
+                      targetColumn?.min,
+                      formatAsAbbreviatedNumber,
+                    )
+              }
+            />
+            <MetricsInfo
+              {...props}
+              name="Max"
+              metakey="max"
+              firstSlot={formatColumnValueWith(
+                baseColumn?.max,
+                formatAsAbbreviatedNumber,
+              )}
+              secondSlot={
+                isTargetNull
+                  ? NO_VALUE
+                  : isTargetUndefined
+                  ? undefined
+                  : formatColumnValueWith(
+                      targetColumn?.max,
+                      formatAsAbbreviatedNumber,
+                    )
+              }
+            />
+          </>
+        )}
         <MetricsInfo
           {...props}
-          name="Min"
-          metakey="min"
-          firstSlot={formatColumnValueWith(
-            baseColumn?.min,
-            formatAsAbbreviatedNumber,
-          )}
+          name="Distinct"
+          metakey="distinct"
+          firstSlot={formatColumnValueWith(baseDistinct, formatNumber)}
           secondSlot={
             isTargetNull
               ? NO_VALUE
-              : isTargetUndefined
-              ? undefined
               : formatColumnValueWith(
-                  targetColumn?.min,
-                  formatAsAbbreviatedNumber,
+                  targetColumn ? targetDistinctOfTotal : baseDistinctOfTotal,
+                  formatIntervalMinMax,
                 )
           }
         />
         <MetricsInfo
           {...props}
-          name="5%"
-          metakey="p5"
-          firstSlot={formatColumnValueWith(
-            baseColumn?.p5,
-            formatAsAbbreviatedNumber,
-          )}
+          name="Duplicates"
+          metakey="duplicates"
+          firstSlot={formatColumnValueWith(baseDuplicates, formatNumber)}
           secondSlot={
             isTargetNull
               ? NO_VALUE
-              : isTargetUndefined
-              ? undefined
               : formatColumnValueWith(
-                  targetColumn?.p5,
-                  formatAsAbbreviatedNumber,
-                )
-          }
-        />
-        <MetricsInfo
-          {...props}
-          name="25%"
-          metakey="p25"
-          firstSlot={formatColumnValueWith(
-            baseColumn?.p25,
-            formatAsAbbreviatedNumber,
-          )}
-          secondSlot={
-            isTargetNull
-              ? NO_VALUE
-              : isTargetUndefined
-              ? undefined
-              : formatColumnValueWith(
-                  targetColumn?.p25,
-                  formatAsAbbreviatedNumber,
-                )
-          }
-        />
-        <MetricsInfo
-          {...props}
-          name="50%"
-          metakey="p50"
-          firstSlot={formatColumnValueWith(
-            baseColumn?.p50,
-            formatAsAbbreviatedNumber,
-          )}
-          secondSlot={
-            isTargetNull
-              ? NO_VALUE
-              : isTargetUndefined
-              ? undefined
-              : formatColumnValueWith(
-                  targetColumn?.p50,
-                  formatAsAbbreviatedNumber,
-                )
-          }
-        />
-        <MetricsInfo
-          {...props}
-          name="75%"
-          metakey="p75"
-          firstSlot={formatColumnValueWith(
-            baseColumn?.p75,
-            formatAsAbbreviatedNumber,
-          )}
-          secondSlot={
-            isTargetNull
-              ? NO_VALUE
-              : isTargetUndefined
-              ? undefined
-              : formatColumnValueWith(
-                  targetColumn?.p75,
-                  formatAsAbbreviatedNumber,
-                )
-          }
-        />
-        <MetricsInfo
-          {...props}
-          name="95%"
-          metakey="p95"
-          firstSlot={formatColumnValueWith(
-            baseColumn?.p95,
-            formatAsAbbreviatedNumber,
-          )}
-          secondSlot={
-            isTargetNull
-              ? NO_VALUE
-              : isTargetUndefined
-              ? undefined
-              : formatColumnValueWith(
-                  targetColumn?.p95,
-                  formatAsAbbreviatedNumber,
-                )
-          }
-        />
-        <MetricsInfo
-          {...props}
-          name="Max"
-          metakey="max"
-          firstSlot={formatColumnValueWith(
-            baseColumn?.max,
-            formatAsAbbreviatedNumber,
-          )}
-          secondSlot={
-            isTargetNull
-              ? NO_VALUE
-              : isTargetUndefined
-              ? undefined
-              : formatColumnValueWith(
-                  targetColumn?.max,
-                  formatAsAbbreviatedNumber,
+                  targetColumn
+                    ? targetDuplicatesOfTotal
+                    : baseDuplicatesOfTotal,
+                  formatIntervalMinMax,
                 )
           }
         />
