+++ static_report/src/components/shared/QuantilesMatrix.tsx
@@ -9,18 +9,18 @@ import {
   Tr,
 } from '@chakra-ui/react';
 import { nanoid } from 'nanoid';
-import { schemaMetaDescriptions } from '../../../sdlc/schema-meta';
-import { ColumnSchema } from '../../../sdlc/single-report-schema';
-import { ZColSchema } from '../../../types';
+import { schemaMetaDescriptions } from '../../sdlc/schema-meta';
+import { ColumnSchema } from '../../sdlc/single-report-schema';
+import { ZColSchema } from '../../types';
 import {
   formatAsAbbreviatedNumber,
   formatColumnValueWith,
-} from '../../../utils/formatters';
+} from '../../utils/formatters';
 
 type Props = {
   columnDatum: ColumnSchema;
 };
-export function QuantilesChart({ columnDatum }: Props) {
+export function QuantilesMatrix({ columnDatum }: Props) {
   ZColSchema.parse(columnDatum);
   const { min, p5, p25, p50, p75, p95, max } = columnDatum;
   const quantileData: {
