+++ static_report/src/pages/SRColumnDetailsPage.tsx
@@ -1,58 +1,20 @@
-import { SearchIcon } from '@chakra-ui/icons';
-import {
-  Box,
-  Divider,
-  Flex,
-  Grid,
-  GridItem,
-  Input,
-  InputGroup,
-  InputLeftElement,
-  Tab,
-  TabList,
-  TabPanel,
-  TabPanels,
-  Tabs,
-  Tag,
-  TagLabel,
-  Text,
-} from '@chakra-ui/react';
-import { useState } from 'react';
-import { useLocation, useRoute } from 'wouter';
-import { CategoricalBarChart } from '../components/shared/Charts/CategoricalBarChart';
-import { FlatStackedBarChart } from '../components/shared/Charts/FlatStackedBarChart';
-import { HistogramChart } from '../components/shared/Charts/HistogramChart';
-import { getDataChart } from '../components/shared/ColumnCard';
-import { ColumnCardDataVisualContainer } from '../components/shared/ColumnCard/ColumnCardDataVisualContainer';
+import { Box, Divider, Flex, Grid, GridItem, Text } from '@chakra-ui/react';
+import { useRoute } from 'wouter';
+import { QuantilesMatrix } from '../components/shared/QuantilesMatrix';
 import { ColumnCardHeader } from '../components/shared/ColumnCard/ColumnCardHeader';
-import { ColumnDetailListItem } from '../components/shared/ColumnDetailListItem';
-import { GeneralTableColumn } from '../components/shared/GeneralTableColumn';
 import { Main } from '../components/shared/Main';
 import { NumericTableColumn } from '../components/shared/NumericTableColumn';
-import { TextNumberTableColumn } from '../components/shared/TextNumberTableColumn';
-import { ColumnSchema, SingleReportSchema } from '../sdlc/single-report-schema';
+import { SingleReportSchema } from '../sdlc/single-report-schema';
 import { formatTitleCase } from '../utils/formatters';
-import { transformCompositionAsFlatStackInput } from '../utils/transformers';
-type ProfilerGenericTypes = ColumnSchema['type'];
+import { FlatBoxPlotChart } from '../components/shared/Charts/FlatBoxPlotChart';
+import { SRColumnDetailsMasterList } from '../components/SingleReport/SRColumnDetailsMasterList';
+import { DataCompositionWidget } from '../components/shared/Widgets/DataCompositionWidget';
+import { ChartTabsWidget } from '../components/shared/Widgets/ChartTabsWidget';
 interface Props {
   data: SingleReportSchema;
 }
 export function SRColumnDetailsPage({ data: { tables } }: Props) {
-  const [filterState, setFilterState] = useState<
-    Map<ProfilerGenericTypes, boolean>
-  >(
-    new Map([
-      ['boolean', true],
-      ['datetime', true],
-      ['integer', true],
-      ['numeric', true],
-      ['other', true],
-      ['string', true],
-    ]),
-  );
-  const [filterString, setFilterString] = useState<string>('');
   const [match, params] = useRoute('/tables/:reportName/columns/:columnName');
-  const [location, setLocation] = useLocation();
 
   if (!params?.columnName) {
     return (
@@ -66,100 +28,25 @@ export function SRColumnDetailsPage({ data: { tables } }: Props) {
 
   const { reportName, columnName } = params;
   const dataColumns = tables[reportName].columns;
-  const columnEntries = Object.entries(dataColumns);
-  const quickFilters = Array.from(filterState.keys());
 
   const columnDatum = dataColumns[columnName];
-  const { type, topk, histogram, trues, falses, total, min, max } = columnDatum;
-  const showGenericTypeComp =
-    type === 'integer' || type === 'numeric' || type === 'string';
-  const dataCompInput = transformCompositionAsFlatStackInput(
-    columnDatum,
-    'static',
-  );
-  const dynamicCompInput = transformCompositionAsFlatStackInput(
-    columnDatum,
-    'dynamic',
-  );
+  const { type } = columnDatum;
 
   return (
     <Main>
       <Flex
         width={'inherit'}
         minHeight="90vh"
+        maxHeight="100vh"
         p={1}
         bg={'gray.200'}
         direction={['column', 'row']}
       >
         {/* Master Area */}
-        <Flex
-          width={['100vw', '40vw']}
-          direction={'column'}
-          py={8}
-          px={8}
-          mr={2}
-          bg={'white'}
-        >
-          <Text as={'h3'} fontWeight={'bold'} mb={3}>
-            Columns ({columnEntries.length})
-          </Text>
-          {/* Search Filter FIXME: Extract to `TextFilterInput` */}
-          <InputGroup my={2}>
-            <InputLeftElement
-              pointerEvents={'none'}
-              children={<SearchIcon color={'gray.300'} />}
-            />
-            <Input
-              type={'text'}
-              placeholder="Find By Column Name"
-              value={filterString}
-              onChange={({ target }) => setFilterString(target.value)}
-            />
-          </InputGroup>
-          {/* Tag Filters FIXME: Extract to `TagFilterBar` */}
-          <Box mb={6}>
-            <Text as={'small'}>Applied Filters:</Text>
-            <Flex alignItems={'center'}>
-              {quickFilters.map((v) => {
-                const itemValue = filterState.get(v);
-                return (
-                  <Tag
-                    key={v}
-                    m={1}
-                    backgroundColor={itemValue ? 'piperider.300' : ''}
-                    onClick={() => {
-                      const newState = new Map(filterState).set(v, !itemValue);
-                      setFilterState(newState);
-                    }}
-                    cursor={'pointer'}
-                  >
-                    <TagLabel color={itemValue ? 'white' : ''} fontSize={'sm'}>
-                      {v}
-                    </TagLabel>
-                  </Tag>
-                );
-              })}
-            </Flex>
-          </Box>
-          {/* QueryList */}
-          {columnEntries
-            .filter(([key, { type }]) => filterState.get(type))
-            .filter(([key]) =>
-              filterString
-                ? key.search(new RegExp(filterString, 'gi')) > -1
-                : true,
-            )
-            .map(([key, value]) => (
-              <ColumnDetailListItem
-                key={key}
-                datum={value}
-                onSelect={(name) => {
-                  setLocation(`/tables/${reportName}/columns/${name}`);
-                }}
-                p={2}
-              />
-            ))}
-        </Flex>
+        <SRColumnDetailsMasterList
+          dataColumns={dataColumns}
+          currentReport={reportName}
+        />
 
         {/* Detail Area */}
         <Grid
@@ -175,112 +62,28 @@ export function SRColumnDetailsPage({ data: { tables } }: Props) {
           </GridItem>
           {/* Data Composition Block */}
           <GridItem p={9} bg={'white'}>
-            {dataCompInput && (
-              <Box mb={6}>
-                <Text fontWeight={'bold'} fontSize={'xl'}>
-                  Data Composition
-                </Text>
-                <Box height={'55px'}>
-                  <FlatStackedBarChart data={dataCompInput} />
-                </Box>
-                <Box mt={6}>
-                  <GeneralTableColumn baseColumn={columnDatum} width={'100%'} />
-                </Box>
-              </Box>
-            )}
-            {showGenericTypeComp && dynamicCompInput && (
-              <Box>
-                <Text fontWeight={'bold'} fontSize={'xl'}>
-                  {formatTitleCase(type)} Composition
-                </Text>
-                <Box height={'55px'}>
-                  <FlatStackedBarChart data={dynamicCompInput} />
-                </Box>
-                <Box mt={6}>
-                  <TextNumberTableColumn
-                    baseColumn={columnDatum}
-                    width={'100%'}
-                  />
-                </Box>
-              </Box>
-            )}
+            <DataCompositionWidget columnDatum={columnDatum} />
           </GridItem>
           {/* Chart Block - toggleable tabs */}
           <GridItem gridRow={'span 1'} minWidth={0} p={9} bg={'white'}>
-            <Box ml={3}>
-              <Text fontWeight={'bold'} fontSize={'xl'}>
-                Visualizations
-              </Text>
-              {/* FIXME: Weird bug when switching from 1+n tab */}
-              <Tabs>
-                <TabList>
-                  {topk && <Tab>Categorical</Tab>}
-                  {histogram && <Tab>Histogram</Tab>}
-                  {trues && falses && <Tab>Boolean</Tab>}
-                  {type === 'other' && <Tab>Other</Tab>}
-                </TabList>
-
-                <TabPanels>
-                  {topk && (
-                    <TabPanel>
-                      <ColumnCardDataVisualContainer
-                        p={0}
-                        title={columnName}
-                        allowModalPopup
-                      >
-                        <CategoricalBarChart data={topk} total={total || 0} />
-                      </ColumnCardDataVisualContainer>
-                    </TabPanel>
-                  )}
-                  {histogram && (
-                    <TabPanel>
-                      <ColumnCardDataVisualContainer
-                        p={0}
-                        title={columnName}
-                        allowModalPopup
-                      >
-                        <HistogramChart
-                          data={{ histogram, min, max, type, total }}
-                        />
-                      </ColumnCardDataVisualContainer>
-                    </TabPanel>
-                  )}
-                  {trues && falses && (
-                    <TabPanel>
-                      <ColumnCardDataVisualContainer
-                        p={0}
-                        title={columnName}
-                        allowModalPopup
-                      >
-                        {getDataChart(columnDatum)}
-                      </ColumnCardDataVisualContainer>
-                    </TabPanel>
-                  )}
-                  {type === 'other' && (
-                    <TabPanel>
-                      <ColumnCardDataVisualContainer p={0} title={columnName}>
-                        {getDataChart(columnDatum)}
-                      </ColumnCardDataVisualContainer>
-                    </TabPanel>
-                  )}
-                </TabPanels>
-              </Tabs>
-            </Box>
+            <ChartTabsWidget columnDatum={columnDatum} />
           </GridItem>
-          {/* Summary Block FIXME: TBD UI layout */}
           <GridItem gridRow={'span 1'} p={9} bg={'white'}>
             <Box>
-              <Text fontWeight={'bold'} fontSize={'xl'}>
-                {formatTitleCase(type)} Statistics
-              </Text>
+              <Text fontSize={'xl'}>{formatTitleCase(type)} Statistics</Text>
               <Divider my={3} />
               <NumericTableColumn baseColumn={columnDatum} width={'100%'} />
             </Box>
           </GridItem>
           {/* Quantiles Block */}
-          <GridItem gridRow={'span 1'} p={9} bg={'red.300'}>
-            Quantiles Block
-          </GridItem>
+          {(type === 'integer' || type === 'numeric') && (
+            <GridItem gridRow={'span 1'} p={9} bg={'white'} minWidth={'0px'}>
+              <Box>
+                <FlatBoxPlotChart columnDatum={columnDatum} />
+              </Box>
+              <QuantilesMatrix columnDatum={columnDatum} />
+            </GridItem>
+          )}
         </Grid>
       </Flex>
     </Main>
